refactor(console-e2e): share the real-instance sign-in and boundary helpers
`signIn` had been copied byte-for-byte into all three real-instance specs, and
the delete-user suite made it three. It is the step every one of them starts
with, so a drift between copies would have the suites signing in differently
while looking identical. `expectNoErrorBoundary` and its heading list came along
for the same reason.

Also corrects a comment that still described the users list as taking the
default page size of 20 after the loader started asking for the maximum.
